feat: Console 다형성 구현
