+++ src/main/java/com/programmers/voucher/view/InputView.java
@@ -1,12 +1,6 @@
 package com.programmers.voucher.view;
 
-import java.util.Scanner;
 
-public class InputView {
-    private final Scanner scanner = new Scanner(System.in);
-
-    public String getRequest() {
-        String userInput = scanner.nextLine();
-        return userInput;
-    }
+public interface InputView {
+    String getRequest();
 }
